backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Cache y colas
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Mercado Pago
mercadopago==2.2.0

# Email
sendgrid==6.11.0

# Rate limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
